requirements-desktop.txt
# Desktop runtime dependency
# Used by: session-manager-desktop.py
pywebview>=5.0
